--- client/src/views/AdminSlidebar.vue
@@ -1,13 +1,0 @@
-<template>
-
-</template>
-
-<script>
-export default {
-name: "AdminSlidebar"
-}
-</script>
-
-<style scoped>
-
-</style>
