+++ server/budget/views.py
@@ -32,41 +32,3 @@ def budgetEdit(request):
         })
         response.status_code = 200
     return response
-
-
-@onlyForLogginedChecker
-def returnUserInfo(request):
-    try:
-        budget = Profile.objects.get(user=request.user).budget
-        response = JsonResponse({
-            "result": "successfully",
-            "totalBudget": budget.totalBudget,
-            "budgetItems": [{"id": x.id, "title": x.title} for x in budget.budgetItem.all()],
-            "families": [{"id": x.id, "title": x.title} for x in Family.objects.filter(users=request.user)]
-        })
-        response.status_code = 200
-    except:
-        response = JsonResponse({
-            "result": "error"
-        })
-        response.status_code = 500
-    return response
-
-
-@onlyForLogginedChecker
-def returnBudgetOfUser(request):
-    body = request.GET
-    try:
-        budget = Profile.objects.get(user=request.user).budget
-        response = JsonResponse({
-            "result": "successfully",
-            "totalBudget": budget.totalBudget,
-            "budgetItems": [[x.title, x.budget] for x in budget.budgetItem.all()]
-        })
-        response.status_code = 200
-    except:
-        response = JsonResponse({
-            "result": "error"
-        })
-        response.status_code = 500
-    return response
