+++ server/templates/base-vue.html
@@ -5,7 +5,7 @@
     <head>
         {% block head %}{{ block.super }}{% endblock %}
         <!-- html-webpack-inject will inject head assets below -->
-    <link href="http://localhost:8080/js/0.js" rel="prefetch"><link href="http://localhost:8080/js/1.js" rel="prefetch"><link href="http://localhost:8080/js/2.js" rel="prefetch"><link href="http://localhost:8080/js/3.js" rel="prefetch"><link href="http://localhost:8080/js/4.js" rel="prefetch"><link href="http://localhost:8080/js/5.js" rel="prefetch"><link href="http://localhost:8080/js/6.js" rel="prefetch"><link href="http://localhost:8080/js/7.js" rel="prefetch"><link href="http://localhost:8080/js/8.js" rel="prefetch"><link href="http://localhost:8080/js/app.js" rel="preload" as="script"><link href="http://localhost:8080/js/chunk-vendors.js" rel="preload" as="script"></head>
+    <link href="http://localhost:8080/js/0.js" rel="prefetch"><link href="http://localhost:8080/js/2.js" rel="prefetch"><link href="http://localhost:8080/js/3.js" rel="prefetch"><link href="http://localhost:8080/js/6.js" rel="prefetch"><link href="http://localhost:8080/js/7.js" rel="prefetch"><link href="http://localhost:8080/js/app.js" rel="preload" as="script"><link href="http://localhost:8080/js/chunk-vendors.js" rel="preload" as="script"></head>
     <body {% block body_tag_attributes %}{{ block.super }}{% endblock %}>
         {% block body %}{{ block.super }}{% endblock %}
         <!-- html-webpack-inject will inject body assets below -->
